Make clippy happy

+++ src/lib/crypto/kimchi_bindings/stubs/src/pasta_fp_plonk_proof.rs
@@ -201,7 +201,7 @@ pub fn caml_pasta_fp_plonk_proof_example_with_lookup(
     let mut runtime_tables_setup = vec![];
     for table_id in 0..num_tables {
         let cfg = RuntimeTableCfg {
-            id: table_id as i32,
+            id: table_id,
             first_column: [8u32, 9, 8, 7, 1].into_iter().map(Into::into).collect(),
         };
         runtime_tables_setup.push(cfg);
